Fixing bug in https://github.com/nazuke/SEOMacroscope/issues/29
Fixed bad error handling in AnalyzeKeywordPresence(), and added new flag for malformed keywords meta tag

+++ SEOMacroscopeSeriesOne/src/MacroscopeForms/MacroscopeDisplay/MacroscopeDisplaySiteOverview/MacroscopeDisplayStructureOverview.cs
@@ -538,7 +538,7 @@ namespace SEOMacroscope
 
           {
             TreeNode[] Leaf = this.tvTreeView.Nodes.Find( "UrlsInternal", true );
-            int Count = DocCollection.CountUrlsInternal();
+            ulong Count = DocCollection.CountUrlsInternal();
             if( Leaf.Length > 0 )
             {
               Leaf[ 0 ].Text = string.Format( "Total Internal URLs: {0}", Count );
@@ -548,7 +548,7 @@ namespace SEOMacroscope
 
           {
             TreeNode[] Leaf = this.tvTreeView.Nodes.Find( "UrlsExternal", true );
-            int Count = DocCollection.CountUrlsExternal();
+            ulong Count = DocCollection.CountUrlsExternal();
             if( Leaf.Length > 0 )
             {
               Leaf[ 0 ].Text = string.Format( "Total External URLs: {0}", Count );
@@ -621,7 +621,7 @@ namespace SEOMacroscope
 
         {
           TreeNode[] Leaf = this.tvTreeView.Nodes.Find( "SitemapsFound", true );
-          int Count = DocCollection.CountUrlsSitemaps();
+          ulong Count = DocCollection.CountUrlsSitemaps();
           if( Leaf.Length > 0 )
           {
             Leaf[ 0 ].Text = string.Format( "Sitemaps Found: {0}", Count );
